fix: all scroll when the gallery is filled with images on mobile
