+++ components/sections/homepage/Gallery.jsx
@@ -21,56 +21,14 @@ export default function Gallery() {
     );
   }, []);
 
-  const getX = (x, width, container) => {
-    let newValue = x - width / 2;
-    if (newValue < 0) {
-      newValue = 0;
-    } else if (x + width >= container) {
-      newValue = container - width;
-    }
-    return newValue;
-  };
-  const getY = (y, height, container) => {
-    let newValue = y - height / 2;
-    if (y < 0) {
-      newValue = 0;
-    } else if (y + height >= container) {
-      newValue = container - height;
-    }
-    return newValue;
-  };
-
-  const handleMouse = (event) => {
-    const randomNumber = Math.floor(Math.random() * images.length);
-    if (coordinates.length < 1) {
-      let x = event.evt.offsetX;
-      let y = event.evt.offsetY;
-      setCoordinates([
-        ...coordinates,
-        {
-          x,
-          y,
-          img: images[randomNumber],
-          w: images[randomNumber].current.clientWidth,
-          h: images[randomNumber].current.clientHeight,
-          containerWidth: event.currentTarget.attrs.container.clientWidth,
-          containerHeight: event.currentTarget.attrs.container.clientHeight,
-          first: true,
-        },
-      ]);
-    } else {
-      if (
-        Math.abs(event.evt.offsetX - coordinates[coordinates.length - 1].x) >
-          100 ||
-        Math.abs(event.evt.offsetY - coordinates[coordinates.length - 1].y) >
-          100
-      ) {
+  const createEventHandler = (distance) => {
+    return (event) => {
+      const randomNumber = Math.floor(Math.random() * images.length);
+      if (coordinates.length < 1) {
         let x = event.evt.offsetX;
         let y = event.evt.offsetY;
-
         setCoordinates([
-          ...coordinates.slice(0, coordinates.length - 1),
-          { ...coordinates[coordinates.length - 1], first: false, img: null },
+          ...coordinates,
           {
             x,
             y,
@@ -82,17 +40,41 @@ export default function Gallery() {
             first: true,
           },
         ]);
+      } else {
+        if (
+          Math.abs(event.evt.offsetX - coordinates[coordinates.length - 1].x) >
+            distance ||
+          Math.abs(event.evt.offsetY - coordinates[coordinates.length - 1].y) >
+            distance
+        ) {
+          let x = event.evt.offsetX;
+          let y = event.evt.offsetY;
+
+          setCoordinates([
+            ...coordinates.slice(0, coordinates.length - 1),
+            { ...coordinates[coordinates.length - 1], first: false, img: null },
+            {
+              x,
+              y,
+              img: images[randomNumber],
+              w: images[randomNumber].current.clientWidth,
+              h: images[randomNumber].current.clientHeight,
+              containerWidth: event.currentTarget.attrs.container.clientWidth,
+              containerHeight: event.currentTarget.attrs.container.clientHeight,
+              first: true,
+            },
+          ]);
+        }
       }
-    }
+    };
   };
+
   return (
     <Wrapper>
       <Canvas
         container={container}
         dimension={dimension}
-        getX={getX}
-        getY={getY}
-        handleMouse={handleMouse}
+        createEventHandler={createEventHandler}
         coordinates={coordinates}
         mobileContainer={mobileContainer}
         mobileDimension={mobileDimension}
@@ -103,7 +85,7 @@ export default function Gallery() {
           return (
             <img
               ref={images[index]}
-              className="max-h-72"
+              className="max-h-60 xl:max-h-72"
               src={url}
               key={index}
               alt=""
